lil ui fixes

+++ Hammer_Time/Assets/Scripts/Tourny/ProvStandings.cs
@@ -66,7 +66,7 @@ public class ProvStandings : MonoBehaviour
 		{
 			if (cm.playerTeamIndex == provRankList[i].team.id)
 			{
-				scrollbar.value = (i - provRankList.Count) / (1f - provRankList.Count);
+				scrollbar.value = (float)(i - provRankList.Count) / (1f - provRankList.Count);
 				standDisplay[i].panel.enabled = true;
 			}
 			else
